refactor(frontend): extract AiAssistantChat component
